Learned @ManyToMany and one/two way insertion
From my study, I found out that, If we are trying to perform save operation:-

A) 2-way insertion will be needed for the side that has mappedBy

B) 1-way insertion will be enough for the side that doesn't have mappedBy.

C) In the @ManyToOne side, we can not use mappedBy. We can only use it the in @OneToMany side.

(See the comments in the LearnSpringOrmApplication file, and the Entity files for reference).